ci (#3)
* Configure static export

* Remove API from build

* Remove static export and add API

* Add static

* Add output location

* Separate API from app

* Update host.json

* Add extension bundle
