+++ .github/workflows/deploy.yml
@@ -21,6 +21,8 @@ jobs:
           azure_static_web_apps_api_token: ${{ secrets.AZURE_STATIC_WEB_APPS_API_TOKEN_GREEN_ISLAND_09F799A00 }}
           repo_token: ${{ secrets.GITHUB_TOKEN }}
           action: "upload"
-          app_location: "/"
+          app_location: "/web"
           api_location: "/api"
-          output_location: ""
+          output_location: "/out"
+        env:
+          IS_STATIC_EXPORT: true
