+++ api/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "sunday-music-school-api",
+  "version": "1.0.0",
+  "private": "true",
+  "scripts": {
+    "build": "tsc",
+    "watch": "tsc -w",
+    "prestart": "npm run build",
+    "start": "func start",
+    "dev": "npm run watch & npm start",
+    "test": "echo \"No tests yet...\""
+  },
+  "dependencies": {
+    "@sendgrid/mail": "^7.7.0"
+  },
+  "devDependencies": {
+    "@azure/functions": "^1.2.3",
+    "@types/node": "^20.1.0",
+    "typescript": "^4.0.0"
+  }
+}
