+++ web/src/components/GradientText.tsx
