+++ web/src/components/NoteRain.tsx
